WIP: test: Extend validateCiliumSvc to consider all cluster IPs
Previously, the validateCiliumSvc(), which checks whether "cilium
service list" is in sync with "kubectl get svc", did not consider all
cluster IPs. It was only comparing against ".spec.ClusterIP", and did
not compare against ".spec.ClusterIPs" which can store a ClusterIP per
family when the DualStack is on.

The limitation was obvious when trying to backport the recent
K8sServices suite refactoring to the v1.11. The CI was failing with:

    Could not find Cilium service with ip fd03::e44e in k8s

Further investigation showed, that the svc IP addr was in
".spec.ClusterIPs", but not in ".spec.ClusterIP".

Signed-off-by: Martynas Pumputis <m@lambda.lt>